refactor(ocale): migrate universes wire code into domains/cloud-v2
Moves the `cloud-v2` `get` and `update` Operation wire code (builders,
parsers, wire types, rate-limit constants, parsed shapes) for the
Universes Resource into a new per-shape sub-tree under
`src/domains/cloud-v2/universes/`, following ADR-022. Rewrites
`src/resources/universes/client.ts` and `index.ts` to import from the
Domain sub-tree, matching the binding pattern already in place for
`client.icon.*` and `client.thumbnails.*` (introduced in #300). Public
surface, wire requests, parsed shapes, and operation keys
(`universes.get`, `universes.update`) are preserved end-to-end;
conformance and integration tests pass without modification.

Closes #301.
