+++ packages/open-cloud/src/domains/cloud-v2/universes/builders.spec.ts
@@ -1,6 +1,6 @@
 import { assert, describe, expect, it } from "vitest";
 
-import { ValidationError } from "../../errors/validation.ts";
+import { ValidationError } from "../../../errors/validation.ts";
 import { buildGetRequest, buildUpdateRequest } from "./builders.ts";
 
 describe(buildGetRequest, () => {
